refactor the api of runtime instantiation
