+++ src/playground.rs
@@ -13,15 +13,10 @@ use egui_code_editor::{CodeEditor, ColorTheme, Syntax};
 use crate::readback::{Element, RunStats};
 use core::time::Duration;
 use par_builtin::import_builtins;
-use par_core::{
-    execution::TokioSpawn,
-    frontend::{
-        CheckedModule, Module, ParseAndCompileError, Type, TypeError, TypeOnHover, compile_runtime,
-        language::GlobalName, lower, parse, process, type_check,
-    },
-    runtime::{Compiled, RuntimeCompilerError, TypedHandle},
-    source::FileName,
-};
+use par_core::{execution::TokioSpawn, frontend::{
+    CheckedModule, Module, ParseAndCompileError, Type, TypeError, TypeOnHover, compile_runtime,
+    language::GlobalName, lower, parse, process, type_check,
+}, runtime, runtime::{Compiled, RuntimeCompilerError, TypedHandle}, source::FileName};
 use std::collections::HashMap;
 use std::fmt::Write;
 use std::time::Instant;
@@ -572,7 +567,7 @@ impl Playground {
                 *cancel_token = Some(token.clone());
 
                 let ty = name_to_ty.get(name).unwrap();
-                let (handle, reducer_future) = tokio.block_on(compiled.start_and_instantiate(name));
+                let (handle, reducer_future) = tokio.block_on(runtime::start_and_instantiate(&compiled, name));
                 let stats: RunStats = Arc::new(Mutex::new(None));
 
                 let ctx = ui.ctx().clone();
